backend/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
flask>=3.0.0
supabase>=2.0.0
